pull Rizal

+++ database/migrations/2024_09_08_050339_create_items_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('code')->unique();
-            $table->unsignedBigInteger('purchase_price');
-            $table->unsignedBigInteger('selling_price');
+            $table->decimal('purchase_price', 15, 2);
+            $table->decimal('selling_price', 15, 2);
             $table->string('unit');
             $table->integer('stock');
             $table->text('description');
